+++ react/Pixel.tsx
@@ -1,5 +1,6 @@
 import { canUseDOM } from 'vtex.render-runtime'
 
 if (canUseDOM) {
+  // window.newtailMedia = window.newtailMedia || {}
   window.newtailMedia.isReady = true
 }