+++ react/components/Banner/types.d.ts
@@ -3,6 +3,10 @@ type NewtailMediaBannerProps = {
   placementNameAdmin?: string
   size?: string
   sizeAdmin?: string
+  sizeMobile?: string
+  sizeMobileAdmin?: string
   quantity?: number
   quantityAdmin?: string
+  categoryName?: string
+  categoryNameAdmin?: string
 }