+++ CHANGELOG.md
@@ -1,21 +1,40 @@
 # Changelog
 
+## [Unreleased]
+
+## [2.0.10] - 2024-08-16
+- improve resquest body - category name override
+
+## [2.0.9] - 2024-08-16
+- Home context with category home when banner ads
+- Ad size for mobile
+- Declare category name for without context page
+
 ## [2.0.8] - 2024-08-05
 - Home context
 - CSS for multiple banners
 
 ## [2.0.7] - 2024-07-10
-- Adding departament context
-- Refactoring same hooks
-- 
-## [2.0.6]  - 2024-07-03
-- Change click event on shelf
+
+## [2.0.6] - 2024-07-09
 
 ## [2.0.5] - 2024-06-26
+
+## [2.0.4] - 2024-06-26
+
+## [2.0.3] - 2024-06-25
+
+## [2.0.2] - 2024-06-21
+- Adicionando Newtail Search
+
+## [2.0.1] - 2024-06-17
+- Change publisher ID
 
 ## [2.0.0] - 2024-06-17
 
 ## [1.0.1] - 2024-05-31
+
+### Added
 - v1.0.0 Update do Newtail configs
 
 ## [0.0.1]